+++ BicycleRental/src/main/java/com/wxh/bicyclerental/mapper/BicycleMapper.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 @Mapper
 public interface BicycleMapper {
-    public Bicycle selectOne(Integer bicycleCode);
+    public Bicycle selectOne(Long bicycleCode);
     public List<Bicycle> select();
     public int insert(Bicycle bicycle);
-    public int delete(Integer bicycleCode);
+    public int delete(Long bicycleCode);
     public int update(Bicycle bicycle);
 }
